--- src/app/login/auth/callback/page.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const page = () => {
-  return <div></div>;
-};
-
-export default page;
