+++ src/features/auth/index.ts
@@ -1,2 +1,3 @@
 export { LoginButton } from './ui/LoginButton';
 export type { ProviderTypes } from './model/types';
+export { useSocialLogin } from './model/hooks/useSocialLogin';
